Extract UCSC Cell Browser Datasets with Attributes.

Extract UCSC Cell Browser Datasets with Attributes.

Usage

ExtractCBDatasets(
  all.samples.df,
  collection = NULL,
  sub.collection = NULL,
  organ = NULL,
  disease = NULL,
  organism = NULL,
  project = NULL,
  fuzzy.match = TRUE,
  cell.num = NULL
)

Arguments

all.samples.df

Dataframe contains all samples metadata, obtained with ShowCBDatasets.

collection

The collection of the datasets, corresponds to shortLabel column of ShowCBDatasets, obtain available values with StatDBAttribute. Default: NULL (without filtering).

sub.collection

The sub-collection of the datasets, corresponds to subLabel column of ShowCBDatasets, obtain available values with StatDBAttribute. Default: NULL (without filtering).

organ

The organ of the datasets, corresponds to body_parts column of ShowCBDatasets, obtain available values with StatDBAttribute. Default: NULL (without filtering).

disease

The disease of the datasets, corresponds to diseases column of ShowCBDatasets, obtain available values with StatDBAttribute. Default: NULL (without filtering).

organism

The specie of the datasets, corresponds to organisms column of ShowCBDatasets, obtain available values with StatDBAttribute. Default: NULL (without filtering).

project

The project of the datasets, corresponds to projects column of ShowCBDatasets, obtain available values with StatDBAttribute. Default: NULL (without filtering).

fuzzy.match

Logical value, whether to perform fuzzy match with provided attribute values. Default: TRUE.

cell.num

Cell number filter. If NULL, no filter; if one value, lower filter; if two values, low and high filter. Deault: NULL(without filtering).

Value

Dataframe contains filtered datasets.

Examples

# # lazy mode, load datasets json files locally
# ucsc.cb.samples = ShowCBDatasets(lazy = TRUE, json.folder = NULL, update = FALSE)
# # cell number is between 1000 and 2000
# hbb.sample.df = ExtractCBDatasets(all.samples.df = ucsc.cb.samples, organ = c("brain", "blood"),
#                                   organism = "Human (H. sapiens)", cell.num = c(1000,2000))